Added new metric, and now write metrics on screen
Writing on screen was less effort than I remember, Monogame has improved
upon XNA its really great, and I'm so glad you all read all my commit
messages.

+++ RoutingPrototype/RouteManager.cs
@@ -36,8 +36,8 @@ namespace RoutingPrototype
         {
             mScreenWidth = screenWidth;
             mScreenHeight = screenHeight;
-            mXResolutionScaler = screenWidth / 800;
-            mYResolutionScaler = screenHeight / 600;
+            mXResolutionScaler = (float)screenWidth / 800f;
+            mYResolutionScaler = (float)screenHeight / 600f;
             mMapWidth = (int)(mScreenWidth * 0.75f);
             mMarkerTexture = markerText;
             mLineTexture = lineText;
